pyproject.toml
[project]
name = "tds-virtual-ta"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fuzzywuzzy>=0.18.0",
    "psutil>=7.0.0",
    "pytesseract>=0.3.13",
    "uvicorn>=0.22.0",
    "python-multipart",
    "fastapi>=0.100.0",
    "pydantic>=2.0.0",
    "pillow>=9.0.0",
    "requests>=2.26.0",
    "python-multipart>=0.0.5",
    "fuzzywuzzy>=0.18.0",
    "beautifulsoup4>=4.10.0",
    "psutil>=7.0.0",
    "pytesseract>=0.3.8",
    "pillow>=9.0.0",
    "io>=1.0.0",
    "base64>=1.0.0",
    "typing>=3.7.0",
    "time>=3.0.0",
    "datetime>=4.0.0",
    "requests>=2.26.0",
    "beautifulsoup4>=4.10.0",
    "json>=2.0.9",
    "logging>=0.5.1",
    "threading>=3.0.0"
]
